-allows build.xml to build for 1.3 in a 1.5 JDK -removes some casts and semicolons that are not needed -adds some missing @return descriptions in javadocs -deprecates tests that are testing deprecated code so they don't show up as warnings
Patch by Brad Clarke.

+++ main/src/net/sourceforge/cruisecontrol/sourcecontrols/P4.java
@@ -349,11 +349,11 @@ public class P4 implements SourceControl {
                 Calendar cal = Calendar.getInstance();
 
                 cal.setTime(lastBuildTime);
-                cal.add(Calendar.MILLISECOND , (int) offset);
+                cal.add(Calendar.MILLISECOND , offset);
                 lastBuildTime = cal.getTime();
 
                 cal.setTime(now);
-                cal.add(Calendar.MILLISECOND , (int) offset);
+                cal.add(Calendar.MILLISECOND , offset);
                 now = cal.getTime();
 
             } catch (IOException ioe) {
